requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-allauth>=0.57.0
django-crispy-forms>=2.0
crispy-tailwind>=0.5.0
django-filter>=23.0
python-dotenv>=1.0.0
Pillow>=10.0.0
celery>=5.3.0
psycopg2-binary>=2.9.6
dj-database-url>=2.0.0
gunicorn>=20.1.0
whitenoise>=6.5.0
redis
